requirements.txt
piexif==1.1.3
pillow==12.2.0
Send2Trash==2.1.0
